Split doorstubs.cpp into logical homes
doorstubs.cpp was a grab-bag of link-time stubs for Player init. Move each
into the subsystem that owns it so the layout mirrors the original source.

- door.cpp/.h (new): hub/exit-door subsystem - Door_* and hub_* globals
  plus StartDoorPositions, Door_FindByIndex, Door_FindByName.
- cheat.cpp: implement Cheat_IsOn and define the Cheat_PowerUpTime /
  ONEPLAYERPOWERUPS globals it references.
- episode.cpp: implement Episode_FindAreaFromFlags over EPISODEDATA/AREADATA.
- players.cpp: relocate PlayerStart and the player/gameobject helper stubs
  (ComplexSockPosition, ChatterSfx, Move_VEHICLE, DrawOffsetCode,
  rtlDynamicEnable, GameObjectNearFloor, GetHoverPosY, Player_HasPurpleForce,
  GAMEPAD_ACTION, CutScenePlayer_Available) beside the PreResetCode/
  Players_InitPositions callers that use them.

Emits the same symbol names as before so the objdiff targets are unchanged.
