+++ src/legoapi/cheat.h
@@ -9,6 +9,7 @@ void Cheat_SetArea(i32 cheat, i32 areaId);
 i32 Cheat_FindByName(char *name);
 u32 Cheats_CheckFlags(u32 flag);
 u32 Cheat_CheckFlags(int cheat_index, u32 flag_mask);
+int Cheat_IsOn(int cheat);
 struct cheatsystem_s {
     CHEAT *cheats;
     int cheats_count;
