+++ src/legoapi/episode.h
@@ -27,4 +27,6 @@ EPISODEDATA *Episodes_ConfigureList(char *file, VARIPTR *bufferStart, VARIPTR *b
 
 i32 Episode_ContainsArea(i32 areaId, i32 *areaIndex);
 
+u32 Episode_FindAreaFromFlags(EPISODEDATA *ep, u32 flags, u32 want);
+
 #endif
